refactor(InputView): 입력 예외 코드 이동
입력 예외 코드를 InputView로 추가
- BridgeMaker에 있던 코드 삭제
