+++ src/App.js
@@ -1,8 +1,10 @@
 const BridgeGame = require("./BridgeGame");
 const { readBridgeSize } = require("./InputView");
+const { printStart } = require("./OutputView");
 
 class App {
   play() {
+    printStart();
     readBridgeSize();
   }
 }
