+++ src/BridgeMaker.js
@@ -2,9 +2,6 @@
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
  */
 
-const BRIDGELENGTH_ERR_MESSAGE =
-  "[ERROR] 다리 길이는 3부터 20 사이의 숫자여야 합니다.";
-
 const BridgeMaker = {
   /**
    * @param {number} size 다리의 길이
@@ -12,29 +9,16 @@ const BridgeMaker = {
    * @return {string[]} 입력받은 길이에 해당하는 다리 모양. 위 칸이면 U, 아래 칸이면 D로 표현해야 한다.
    */
   makeBridge(size, generateRandomNumber) {
-    bridgeLenValidator(size);
-    [...Array(parseInt(size)).keys()].map((bridgeLenNum) => {});
-    const number = generateRandomNumber();
+    let bridge = "";
+    [...Array(parseInt(size)).keys()].map((bridgeLenNum) => {
+      const number = generateRandomNumber();
+      console.log(number);
+    });
 
     return;
   },
 };
 
-const bridgeLenValidator = (bridgeLen) => {
-  isInRange(bridgeLen);
-  isIntNumber(bridgeLen);
-};
-
-const isInRange = (bridgeLen) => {
-  if (bridgeLen > 20 || bridgeLen < 3) {
-    throw new Error(BRIDGELENGTH_ERR_MESSAGE);
-  }
-};
-
-const isIntNumber = (bridgeLen) => {
-  if (isNaN(parseInt(bridgeLen)) || !Number.isInteger(parseFloat(bridgeLen))) {
-    throw new Error(BRIDGELENGTH_ERR_MESSAGE);
-  }
-};
+const genarateBridgeLine = () => {};
 
 module.exports = BridgeMaker;
