+++ src/InputView.js
@@ -4,8 +4,9 @@ const BridgeRandomNumberGenerator = require("./BridgeRandomNumberGenerator");
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
  */
-
-const INPUT_BRIDGE_LEN_STR = "다리의 길이를 입력해주세요.";
+const BRIDGELENGTH_ERR_MESSAGE =
+  "[ERROR] 다리 길이는 3부터 20 사이의 숫자여야 합니다.";
+const INPUT_BRIDGE_LEN_STR = "다리의 길이를 입력해주세요.\n";
 
 const InputView = {
   /**
@@ -13,6 +14,7 @@ const InputView = {
    */
   readBridgeSize() {
     MissionUtils.Console.readLine(INPUT_BRIDGE_LEN_STR, (bridgeLen) => {
+      bridgeLenValidator(bridgeLen);
       BridgeMaker.makeBridge(bridgeLen, BridgeRandomNumberGenerator.generate);
     });
   },
@@ -26,6 +28,23 @@ const InputView = {
    * 사용자가 게임을 다시 시도할지 종료할지 여부를 입력받는다.
    */
   readGameCommand() {},
+};
+
+const bridgeLenValidator = (bridgeLen) => {
+  isInRange(bridgeLen);
+  isIntNumber(bridgeLen);
+};
+
+const isInRange = (bridgeLen) => {
+  if (bridgeLen > 20 || bridgeLen < 3) {
+    throw new Error(BRIDGELENGTH_ERR_MESSAGE);
+  }
+};
+
+const isIntNumber = (bridgeLen) => {
+  if (isNaN(parseInt(bridgeLen)) || !Number.isInteger(parseFloat(bridgeLen))) {
+    throw new Error(BRIDGELENGTH_ERR_MESSAGE);
+  }
 };
 
 module.exports = InputView;
